+++ client/src/components/footer.js
@@ -4,7 +4,7 @@ const Footer = () => {
     return (
         <footer className="bg-white shadow bg-gray-50 dark:bg-gray-800 mt-auto">
             <div className="w-full mx-auto max-w-screen-xl p-4 md:flex md:items-center md:justify-center">
-                <span className="text-sm text-gray-500 sm:text-center dark:text-gray-400">© 2023 <a href="https://github.com/Ruben-JR" class="hover:underline">Ruben de Pina</a>.</span>
+                <span className="text-sm text-gray-500 sm:text-center dark:text-gray-400">© 2023 <a href="https://github.com/Ruben-JR" className="hover:underline">Ruben de Pina</a>.</span>
             </div>
         </footer>
     );
